line chart: make threejs coordinator match DOM size
Previously, we had a camera coordinte system where we had a fixed camera
view point. This was great because irrespective of the actual DOM size,
we could keep the objects position static. However, this approach has a
major problem. Now that we want to render shapes like triangle and
circle, geometry that is defined in the camera coordinate system has to
account for mismatch in camera coordinate aspect ratio vs. DOM camera
coordinte aspect ratio.

We want circle that is not an oval in browser pixel location. To do
that, when creating a geometry in the camera coordinates, we need to
account for discrepancy in aspect ratio, which generally is harder than
just matching the camera view point to that of the DOM size. As such,
this change removes the camera coordinate system separate from DOM but
supports the y-axis flipping which is the only remaining difference in
the coordinate systems we support.