rust: bind server to `::1` by default (#4440)
Summary:
Previously, the data server defaulted to host `::0`, the IPv6 wildcard
address and analogue of `0.0.0.0`, ~a.k.a. `--bind_all`. But we can
safely restrict this to just `::1`, the loopback address and analogue of
`127.0.0.1`, allowing localhost connections only. @psybuzz reports that
macOS shows a security dialog about incoming connections to the data
server, and that this fixes it.

Test Plan:
Running `tensorboard --load_fast` still works fine, since TensorBoard is
connecting to a data server running on the same machine.

wchargin-branch: rust-default-bind-loopback