ci: collapse some build jobs (#4448)
Summary:
Previously, we had separate linting jobs for Python, docs, build, and
miscellaneous checks. Now that we have our main heavyweight build on
GitHub Actions as well (#2953), we spend more time in actions overall,
and queuing behavior becomes more important. This patch collapses some
of those jobs together.

I haven’t actually run into any issues here. This is just an ounce of
prevention.

The advantage is that we spin up fewer VMs and thus use our resources
more efficiently. The disadvantage is that we fail faster: e.g., an
error in Python style will now prevent ipynb lints from running. This
seems fine to me, since the checks are still fast and we don’t usually
have large Python changes and large docs changes in the same PR. If it’s
a problem, we could probably rig it to run all steps of a job and fail
only at the end, but I haven’t looked into that.

We keep Rust and JavaScript lints in their own jobs because they need
some toolchain setup that can be expensive. We keep flake8 in its own
job because it blocks running the build (to catch silly syntax/name
errors) and thus is on the critical path. We could in principle further
merge `lint-python-yaml-docs` and `lint-misc`.

Test Plan:
Note that this PR spawns a workflow with 12 jobs instead of 15.

wchargin-branch: ci-collapse-jobs