graph: use data provider if no multiplexer available
Summary:
The graph plugin still defaults to the multiplexer code path, since not
all functionality is implemented in the data provide code path (only run
graphs are implemented). But when a data provider is available and a
multiplexer is not, it should still try to use the data provider. Now,
it does so. This fixes a failure in `is_active`, which previously tried
to call `PluginRunToTagToContent` on `None`.

Test Plan:
Run TensorBoard with `--load_fast` (causing it to provide a gRPC data
provider but no multiplexer) and launch the web server to trigger a
request to `/plugins_listing`. Note that this no longer prints an error
that “is_active() for graphs failed”.

wchargin-branch: graph-data-provider-fallback
wchargin-source: 692b4fc2f444f6350756d27de17bc0b442cd0684