cleanup: remove unused tf_web_test
tf_web_test is a build rule specialized to test Polymer 1 components.
While we want to have tests, we are (1) making great strides rewriting
critical pieces in Angular (2) not writing much code in Polymer and (3)
hard to maintain tf_web_test which is largely built around
tf_web_library and rules_closure.

Since we are not actively working to rewrite the tests, it would make
sense to remove these deadcodes.