line chart: make sure tooltip repositions when scrolled (#4463)
The new line chart uses cdk overlay to render the tooltip. It currently
has three strategies it can take when scrolled in the scroll container.
Strategies aside, we need a way for the overlay to be able to detect a
scroll event in a scroll container and adding `cdkScrollable` is one way
to register a scrollable container to the cdk library.