Provide and consume FeatureFlagState StoreConfig, allowing other Tensorboards to override default feature flag values. (#4458)
Motivation for features / changes:

Some instances of TensorBoard (notably, TensorBoard.dev), might want to provide different default
values for feature flags.

Technical description of changes:

Define injection token that allows us to provide and consume a StoreConfig. StoreConfig
allows an initialState value to be specified for the state.

Provide a fallback StoreConfig that simply defines initialState to be. . . the initialState we already use.