line chart: introduce formatter (#4474)
Different use cases require you to format numbers a bit differently. For
instance, when viewing line chart in the time series dashboard, with the
wall time x axis type, you want to represent the numbers in time. This
change formalizes a notion of formatter that tooltip and axis can reuse.