graph: use DataProvider-compatible APIs (#4470)
Summary:
Events with top-level `tagged_run_metadata` are now transformed at read
time into blob sequence summaries. As a result, the graph plugin can get
all its data via tensor summaries. This also includes replacing calls to
`multiplexer.Graph` with reads of the `__run_graph__` tensor.

Since there’s an existing `PLUGIN_NAME_RUN_METADATA`, I had initially
hoped to use that for the summaries. But the graphs plugin actually
expects that data to include a graph (even though there is a separate
plugin called `PLUGIN_NAME_RUN_METADATA_WITH_GRAPH`) and there’s no way
to tell it apart from just the metadata. Thus, we implement this via a
new plugin name, which means that there are few structural changes to
the graph plugin code.

As written, the code still uses the multiplexer, but is set up to make
it easy to read from the data provider instead. We make that change in a
follow-up PR (see #4473).

Test Plan:
All data from the graphs demo (added in #4469) still works. Grepping for
`_multiplexer` in the graphs plugin code shows that the only calls are
to `PluginRunToTagToContent` and `Tensors`, both of which have
equivalents in the data provider API.

wchargin-branch: graph-data-provider-compatible