Pass FeatureFlags to CUSTOM_ELEMENT dashboards; use in Scalars dashboard for inColab and scalarsBatchSize. (#4466)
Motivation for features / changes

We want to be able to use the FeatureFlags mechanism to modify behavior in dashboards.
The Scalars dashboard, for example, could use two FeatureFlags:

  * isColab, which maps to the tensorboardColab query parameter. The Scalars dashboard already independently
reads this query parameter but it is more consistent to get it from FeatureFlags.
  * scalarsBatchSize, which is new functionality. TensorBoard.dev will use FeatureFlags to modify the /scalars_multirun
batch size.

Technical description of changes

  * plugins_component on Angular side, which manages dashboard creation, now waits for FeatureFlags to be
loaded before loading dashboards. It passes the featureFlags property of each dashboard, which can choose to
use or ignore it as a polymer property.
  * introduce the scalarsBatchSize feature flag and map it to the query parameter of the same name.
  * scalars dashboard code uses the featureFlags property to read isColab and scalarsBatchSize feature flags
and uses the values to modify behavior.