line chart: transfer data with 64 bit precision (#4475)
We premature optimized thinking that we should allocate less memory
using Float32Array where possible instead of Float64Array. The thing
is, with 32 bit precision, it is not possible to represent the date
serialized to a number.

This changes the data transfer layer only. We will diagnose and change
other places where we use Float32Array as needed (for instance, WebGL
uses 32 bit precision and threejs internally uses Float32Array a lot.
Having more precision does not help in all cases).