ci: persist Rust caches for Pip package build
Summary:
The `build-data-server-pip` job runs on two platforms on every push, and
each run takes about 5 minutes. The vast majority of this time (85%) is
spent building all dependencies from scratch. Since [macOS minutes are
priced at 10× the rate of Linux minutes][pricing] and this is our only
macOS job, we may as well spend a modicum of effort to improve this.

[pricing]: https://docs.github.com/en/free-pro-team@latest/github/setting-up-and-managing-billing-and-payments-on-github/about-billing-for-github-actions#about-billing-for-github-actions

The docs at <https://github.com/actions/cache> say that a repository can
have up to 5 GB of caches. The cache for our other Rust job is ~200 MB,
and we don’t have any other caches, so adding two more of those caches
seems okay. The caches are also invalidated infrequently: only when we
update Rust dependencies or update the workflow file itself.

Adding a cache should speed up these builds. We already do this for the
Rust lint job, and it works well. This is a separate cache because the
job builds in release mode rather than debug mode.

Rust builds generally have good hermeticity properties, so this should
be fairly safe. If we want to be extra safe, we could turn this off for
builds that will be published in an actual stable release.

Test Plan:
Check that the second run of the checks for this PR spends a negligible
amount of time in the “Build” and “Test” phases.

wchargin-branch: ci-cache-rust-pip
wchargin-source: 631d316bed4c17387901b166c6d8d28a6f8c8b7f