line chart: temporal axis
This change treats time axis a bit specially and hardcodes logic for
rendering times. Time ticks, unlike simple numbers, can result in
massive confusion if we encode one to fit into 50px. For instance, if I
were representing 2020/12/31 12:31:12.442 as "12.442", you lose all the
important date, hour, and minute. We are definitely using `<title>` so
user can see more complete information but we lose the "glance-ability".

To mitigate the issue (not perfectly solving the issue), show a major
tick only for date which shows a bit higher level information (datetime)
when certain conditions are right. This business logic definitely is not
perfect but it is a great start and found it to be helpful.

I find special treatment for time scale encoded in line_chart_v2 a bit
unfortunate but it is the easiest way out for now.