py3: remove `argparse_util` helper (#4493)
Summary:
In #2793, we added a helper to ensure that you can run `tensorboard`
either with or without an explicit subcommand. This works by default in
Python 3; the helper was only needed for Python 2. Thus, we remove it.

Part of #4488.

Test Plan:
The test plan from #2793 still passes. Instead of patching in a `demo`
subcommand, you can now use `tensorboard dev list [--json]`.

wchargin-branch: py3-no-argparseutil