Add a lengths argument to tensorboard.summary.audio.
This change adds a lengths keyword argument to tensorboard.summary.audio, which
allows the caller to indicate the per-batch-item lengths of each signal in the
dense input tensor.

Motivation: Audio summaries have always assumed that examples in the dense
tensor shaped [batch_size, num_frames, num_channels] all have the same
length. This is frequently not the case when building audio models, such as
text-to-speech models, where the output length of the model varies per batch
item. When ragged tensors of substantially different shapes are batched
together, this can be wasteful. Additionally, the padding after the end of each
record is included in the resulting WAV file, which is misleading because these
samples were not present in the source signal. The consumer of the audio summary
has no way to determine what the original true lengths of each example were.

TESTED:
- unit test