py3: use `html.escape` unconditionally (#4515)
Summary:
The `html.escape` function superseded `cgi.escape` in Python 3.2, prior
to which the `html` module did not exist. We now unconditionally import
`html`, without falling back to `cgi`.

Part of #4488.

Test Plan:
Tested the Pip package in Jupyter. When rebased on top of #4507, which
removes the `unittest.mock` backport, a `git grep -A1 '^try:'` shows
that the remaining top-level conditional statements are okay: a bunch of
TensorFlow compatibility incantations for summaries, and one conditional
import of `botocore`.

wchargin-branch: py3-html-escape