graph: speed up "Normalizing names" step (#4496)
TensorBoard graph has a method of extracting out output tensor key from
the input names. This method currently relies on regex matches which
usually is fast but given that inputName length gets longer (deep
name_scopes) with many nodes make it a hotspot and using the regex only
when we absolutely need it (i.e., when the name contains ':') reduced
the normalization time from 80s to 450ms.