timeseries: stamp scalar only when first visible
In the new gpu line chart, we have an option to use Worker process to
render the line chart but its initialization can be heavy as it requires
to download the worker resources (or even parse script and create a
process). To make the AFT better, we only stamp and create line chart
component when it is actually visible on the screen.