rust: build universal data server with no binary (#4535)
Summary:
The `tensorboard-data-server` package now supports a universal build
configuration. When built in this configuration, the package does not
actually contain a binary, and returns `None` when interrogated for its
path. But it can be installed on any platform, so it serves as a
fallback. Thus, `tensorboard` can depend on `tensorboard-data-server`
while maintaining install-time compatibility on all platforms.

Closes #4533; see that issue for details.

Test Plan:
Running `build.py` with `--universal` successfully creates a Pip package
called `tensorboard_data_server-0.1.0a0-py3-none-any.whl` that contains
no binary. Running TensorBoard with `--load_fast` prints the new error
message and exits. When re-installing the native wheel, TensorBoard can
still find it. The existing codepaths for a bundled server and a server
taken from environment variable also still work.

Uploaded the three wheels built in CI to Test PyPI as `0.1.0a20210111`,
in the order (1) macOS (2) universal (3) Linux, waiting a minute or so
in between each. When installing a patched TensorBoard wheel with a new
`tensorboard-data-server` dep, it behaved as expected per #4533.

wchargin-branch: rust-universal-pip