line chart: set cache when object is returned
There was a bug where when we tried to remove a line while the data is
not yet fetched, the line chart threw a JSE for trying to access
`null.obj3d` in the remove object function. This happened since renderer
can, in some case, early exit by returning null and we populated the
cache with null values. With this change, we disallow adding null value
to the cache with better typing. We previously had the bug because it
was typed to take `any` which includes `null`. Now, we type it to extend
an object which is an okay constraint since it is an Object drawer.