core: cache on JS resources when hash is present (#4558)
Previously, TensorBoard never cached the resource even when its contents
never change in a given pip package. While this is tolerable in the
local instance setting, it is not the best especially when we start to
use worker where the resources would have to be fetched per worker
instance,

This change enables caching for one day (for now) when JavaScript
resources with `_file_hash` query parameter is requested. While this
approach works for any resources, there is no strong reason to do so and
we will start with restrictive caching at first. In the future, we may
apply the query parameter to more resources and enable caching of SVG
and other resources.