line chart: improve color opacity interpolation (#4555)
Previously, we were using d3.color.brighter but its algorithm created
a bit too salient color for certain colors (low luminence like our slate
color) while being light enough for others.

To alleviate this unevenness in brightening, we decided to use the color
interpolation instead. It results in more visually correct result.

Note: changed opacity from 0.4 to 0.25. It is 0.2 in vz-line-chart but
it looked a bit too faint in GPU line chart so we up-ed it by 0.05