rust: pass summary metadata to `into_blob_sequence`
Summary:
For some data types, data-compat conversions of _values_ depend on the
time series _metadata_. For instance, we need to adjust the shapes of
audio tensors to remove labels, and we need to update graph sub-plugins
(Keras models, etc.) to convert scalars to tensors. This patch updates
the `EventValue::into_blob_sequence` interface to require a reference to
the initial summary metadata so that we can add support for such data
types. It’s not yet used.

Test Plan:
Existing unit tests suffice; for an end-to-end test the graphs and
images plugins still work.

wchargin-branch: rust-blob-sequence-metadata
wchargin-source: 318204dbc3eaedf339f13b4d166b9b823248a6af