line chart: update only visible charts (#4540)
In the timeseries dashboard, we keep track of cards that are in the
view. With this information, we can opt out of updating the charts that
are outside of the viewport and save CPU cycles.

This change propagates the visibility information via `enableUpdate`
flag and make sure to apply changes only when the card is visible in the
viewport.