ng: dispatch coreLoaded after router is initialized (#4430)
Only after the router is initialized, we can make sure we are at the
correct routes only at which point we should start to initialize and
make requests. Otherwise, we can make requests at the wrong paths which
may/may-not be correct.

With this change, we fire the coreLoaded after router is initialized
which is confirmed by checking for non-nullness of the activeRoute.