Introduce a selector providing a runId to run map (#4609)
Existing ngrx selectors in the Run feature store offer ways to get a `Run`
from its associated `runId`, and a list of `Run`s for a specific
`experimentId`. However, it would be convenient for consumers who
wish to convert a batch of `runId`s into a batch of `Run`s without having
to do a combination of multiple `getRun()` calls.

This introduces a convenience selector that returns a map that
consumers can use for easy lookup.