grpc_util: extract channel creds parsing and logic (#4601)
Summary:
A new `grpc_util.ChannelCredsType` enum represents the user-facing
variants for channel credentials. Flag parsing and consumption now
delegate to that enum rather than hand-enumerating all cases. The goal
is to enable defining a parallel flag to main TensorBoard to configure
the connection to the data server.

We define one method to create both channel creds and options so that
callers are less likely to get credentials but forget to set options,
which would cause `SSL_DEV` to behave incorrectly.

Test Plan:
Unit tests included. As end-to-end tests:

  - `tensorboard dev list` still works
  - `tensorboard dev --grpc_creds_type=ssl list` also still works
  - `tensorboard dev --grpc_creds_type=local list` successfully parses
    and then fails to connect to addresses, which is expected
  - `tensorboard dev --grpc_creds_type=zzz list` fails to parse with
    “invalid ChannelCredsType value: 'zzz'”

wchargin-branch: grpc-channel-creds-util