line chart: put fades around start&end of text (#4614)
The GPU line chart does not account for space when creating tick string.
This can of course lead to clipping which is kind of fine except it can
sometimes lead to a confusing chart when the substring can make a lot of
sense. While this is mitigated by `<title>` and tooltips, we lose the
glance-ability.

To give user a hint that the tick string might be clipped, we now use
faded mask on the tick strings.