feature: differentiate overridden flags from default (#4612)
Currently, TensorBoard reflects the feature flags as is to the URL. This
means, we cannot differentiate overrides from defaults and makes
changing the defaults harder.

For instance, soon, we want to change fastChart to default true but, in
that case, we don't want the URL to say anything about fastChart. Also,
in that world, we would like to support fastChart=false which should
remain in the URL since it is an explicit override user has requested.

This change, to prevent breakage internally, types defaultFlags and
flagOverrides as optional while keeping the features as required.