line chart: fill the origin with a rectangle (#4615)
We are using CSS `grid` to render the new line chart and this resulted
in an awkward missing pixel at the "origin". To illustrate this with
simple diagram:
```
          |
          |
          |
missing -> ------
```

With this change, we are now rendering the missing origin rectangle with
a span.

About seemingly unrelated change: when using SVG path to render the axis
lines, it was not as crisp as a border and resulted in noticeable visual
jank with the rectangle.