program: restore default assets zip provider
Summary:
This patch reverts the backward-incompatible changes to `program` made
in #4588, while retaining the new functionality in that change.
Specifically, we restore the calling convention of `TensorBoard`
(reordering the parameters was a breaking change, as was removing the
default value). For clarity, the data dependency to `webfiles.zip` lives
in a new Python module, `//tensorboard:assets_lib`. The `program` module
late-imports this only if no assets zip provider is given, in which case
the caller is expected to have a build dep on `:assets_lib`. The main
`//tensorboard` binary does so; `//tensorboard:dev` does not.

Test Plan:
Tested `//tensorboard`, `//tensorboard:dev`, and the Pip package in a
new virtualenv.

wchargin-branch: program-default-assets
wchargin-source: b7dada87af060dc3ffbc5dd2ffee63652f1c1a23