line chart: fix line chart not updating bug
When the line chart is initialized but never resized, currently our
`domDimension` and `viewBox` which is updated lazily after content is
initialized is not reflected to the template and their children. This is
because we use OnPush change detector and we never explicitly ask
Angular to detect changes.

This bug manifests especially when you spam on the tag filter and get
the timing right.