infra: FE development environment (#4524)
With this change, we introduce a new Bazel rule, `tf_dev_js_binary`. It
creates a JavaScript bundle like `tf_js_binary` but instead of being
smart about module dependencies, it simply concatenates all JavaScript
(e.g., compiled TypeScript) sources in the Bazel dependency graph.

This greatly reduces the time we spent for bundling resources and makes
development easier. For instance, making a template change to the run
selector takes <5s of incremental build time (concat + Vulcanization).
It used to take about 18s before.

Do note that this change is imperfect. For instance, we really do not
have to restart the Python webserver when a web resources change as long
as we know to load the new version from the .runfiles. That change would
require a different strategy in loading an asset in the core_plugin
which is out of scope of this change. Another optimization opportunity
is to apply tf_dev_js_binary to every place where we use tf_js_binary.
While simple, since it adds noise, it is postponed to a subsequent PRs.

Lastly, do note that this makes a bundle larger during development. For
example, index.js: 4.0 MiB gzipped (18.7 MiB uncompressed) to 5.3 MiB
(25.8 MiB uncompressed)