line chart: auto reset viewBox when scaleType (#4625)
When user manually changes the viewBox (by scrolling or dragging), we
tend to honor the fixed viewBox even when data/visibility/etc...
changes. While this is reasonable in most cases, when scaleType changes,
it can lead to an illegal domain (e.g., log scale cannot render viewBox
that is negative) which is not great for UX.

As such. this change resets and overwrites the fixed viewBox when
user changes the viewBox.