infra: use terser on tf_js_binary (#4631)
Currently, TensorBoard only uses rollup to bundle and that is it. While
rollup provides some optimization such as tree shaking, it does not do
any variable mangling and useless thing stripping. This change adds
terse to tf_js_bianry which we use for production JS binary.

With the change, our built binary becomes significantly smaller which
should help with download and parse time. We also strip out sourcemap
which contributes to the bloated binary size.