line chart: adjust tooltip positions (#4626)
For line chart, we are using Angular overlay to create, adjust, and
position the tooltip. The overlay supports a position preference
configuration which tries to fit the overlay in the available viewport
space in the order provided in the preference.

Problem: our configuration was a bit faulty and offset[X,Y] were using
wrong signs causing tooltip to overlap with the chart (by 5px) when the
tooltip was placed above the chart. Instead of pushing the tooltip down
by 5px, we meant to place 5px above.

Fix: reviewed the configuration again for their signage and adjusted the
amount for better UX.