types: set `defaultFlags` as required property (#4641)
We are in process of upgrading our FeatureFlag store to divide default
flag vs. overrides. Previously, everything was mixed into `features`
making certain UX impossible.

Because there is a corresponding code internally, we cannot just switch
immediately so this change is the 3rd of 5th step to safely migrate all
use caess.