rust: add `reqwest` dependency (#4643)
Summary:
The [`reqwest`] crate provides a high-level HTTP interface, similar in
spirit to Python’s `requests` package. It has built-in support for JSON
serialization and deserialization via `serde` (which we already use) and
is based on the `hyper` stack (which we also already use), so it should
fit in nicely. We’ll use it to make requests to GCS.

[`reqwest`]: https://crates.io/crates/reqwest

Test Plan:
It builds: `bazel build //third_party/rust:reqwest`.

wchargin-branch: rust-dep-reqwest