rust: extract `Logdir` trait for other filesystems (#4636)
Summary:
Previously, `LogdirLoader` and `RunLoader` were both hard-tied to the
native OS filesystem, via `walkdir::WalkDir` and `std::fs::File`,
respectively. This patch introduces a `Logdir` trait that abstracts over
listing and reading event files. A `DiskLogdir` implementation recovers
the current behavior (mostly; see below), and we can slot in adapters
for more filesystems as desired.

At this point, it’s convenient to drop the hyper-precise semantics
around non-UTF-8 run name collisions, which only occur if you have two
runs whose names are invalid Unicode and are equal after lossy
conversion. We originally handled this precisely because Rust made it
easy to do so, but it’s never come up as an issue in the real world.
It’s no longer quite so convenient to handle, so we cull the complexity.

Test Plan:
Existing unit tests suffice, and a `--load_fast` end-to-end test still
checks out.

wchargin-branch: rust-logdir-trait