Reland "Disable terser on polymer bundle in dev assets mode (#4651)" (#4654)
Re-lands the change in #4651 that was reverted in #4655.
The original change used the `compile` option on `tf_js_binary`
as a signal to skip the Terser minification step. However, this
broke internally because `compile=False` has special semantics
in Google.

Now, the reland does the same minification skip when a new
rule attribute `dev_mode_only=True` is present. Internally,
this attribute is ignored.

Googlers, see test sync cl/355919112.

Before this is landed, cl/355772570 must be submitted first,
to fix syncing issues.