line chart: render unsafe num in log scale
Previously, due to value clipping at the log scale, we were rendering
unsafe value in log scale (non-positive number) as minimum range value.
This change returns `NaN` for correctness and render those values as
triangles.

Note that we are drawing trapezoid for a long seris of NaN as large
amount of THREE.js object causes the line chart to spend ~100ms of
render time per frame.